requirements.txt
scipy
networkx
numpy
yt_dlp
spotipy 
bs4 
youtube_dl 
python-dotenv 
discord.py 
PyNaCl 
ffmpeg
mysql-connector-python==8.0.26 
dblpy==0.4.0 
youtube-search-python==1.4.9 
nltk
wavelink
utility
spotify
Word2Vec
gensim
lavalink
disnake